+++ lib/router/router.dart
@@ -1,12 +1,18 @@
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 
-import '../services/settings/user_notifier.dart';
+import '../services/auth_user_notifier.dart';
+import '../services/user_notifier.dart';
 import 'redirect.dart';
 import 'routes.dart';
 
+final _listenable = Listenable.merge([
+  authUserNotifier,
+  userNotifier,
+]);
+
 final router = GoRouter(
-  refreshListenable: userNotifier,
+  refreshListenable: _listenable,
   redirect: (BuildContext context, GoRouterState state) {
     return guardRedirect(state.uri);
   },