+++ lib/router/routes.dart
@@ -3,7 +3,9 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 
+import '../screens/games/games_screen.dart';
 import '../screens/login/login.dart';
+import '../screens/set_up_profile/set_up_profile_screen.dart';
 import '../screens/settings/settings_screen.dart';
 import '../screens/sign_up/sign_up.dart';
 import '../screens/verify_email/verify_email_screen.dart';
@@ -60,6 +62,16 @@ class VerifyEmailRoute extends GoRouteData {
   }
 }
 
+@TypedGoRoute<SetUpProfileRoute>(path: '/set-up-profile')
+class SetUpProfileRoute extends GoRouteData {
+  const SetUpProfileRoute();
+
+  @override
+  Widget build(BuildContext context, GoRouterState state) {
+    return const SetUpProfileScreen();
+  }
+}
+
 @TypedGoRoute<WaitingForApprovalRoute>(path: '/waiting-for-approval')
 class WaitingForApprovalRoute extends GoRouteData {
   const WaitingForApprovalRoute();
@@ -77,5 +89,15 @@ class SettingsRoute extends GoRouteData {
   @override
   Widget build(BuildContext context, GoRouterState state) {
     return const SettingsScreen();
+  }
+}
+
+@TypedGoRoute<GamesRoute>(path: '/games')
+class GamesRoute extends GoRouteData {
+  const GamesRoute();
+
+  @override
+  Widget build(BuildContext context, GoRouterState state) {
+    return const GamesScreen();
   }
 }