+++ lib/screens/waiting_for_approval/waiting_for_approval_screen.dart
@@ -1,10 +1,17 @@
 import 'package:flutter/material.dart';
 
+import '../settings/settings_button.dart';
+
 class WaitingForApprovalScreen extends StatelessWidget {
   const WaitingForApprovalScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return const Placeholder();
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('Waiting for approval'),
+        actions: const [SettingsButton()],
+      ),
+    );
   }
 }